backend/requirements.txt
flask >= 2.2.2
flask_session >= 0.5.0
Pillow >= 9.2.0
